Used mocking for Throwable methods instead of skip
